+++ src/apps/ScopeChangeRequest/Components/Form/Inputs/ScopeChangeReferences/ScopeChangeReferences.tsx
@@ -4,7 +4,10 @@ import { SearchReferences } from '../../../SearchReferences/SearchReferences';
 
 export const ScopeChangeReferences = (): JSX.Element => {
     const { updateAtom, useAtomState } = scopeChangeFormAtomApi;
-    const onChange = (newList: TypedSelectOption[]) => updateAtom({ references: newList });
+    const onChange = (newList: TypedSelectOption[]) =>
+        updateAtom({
+            references: newList.filter((v, i, a) => a.map((s) => s.value).indexOf(v.value) === i),
+        });
 
     const references = useAtomState(({ references }) => references ?? []);