+++ src/apps/ScopeChangeRequest/api/PCS/getTagById.ts
@@ -1,11 +1,16 @@
-import { httpClient } from '../../../../Core/Client/Functions/HttpClient';
+import { httpClient } from '@equinor/lighthouse-portal-client';
 import { Tag, TagWrapper } from '../../types/ProCoSys/Tag';
 
-export async function getTagById(plantId: string, tagId: number): Promise<Tag> {
+export async function getTagById(
+    plantId: string,
+    tagId: number,
+    signal?: AbortSignal
+): Promise<Tag> {
     const { procosys } = httpClient();
 
     const res = await procosys.fetch(
-        `api/Tag?plantId=${encodeURI(plantId)}&tagId=${tagId}&api-version=4.1`
+        `api/Tag?plantId=${encodeURI(plantId)}&tagId=${tagId}&api-version=4.1`,
+        { signal }
     );
 
     if (!res.ok) {