+++ src/apps/ScopeChangeRequest/hooks/mutations/useWorkflowSigning.tsx
@@ -45,9 +45,11 @@ export function useWorkflowSigning({
         }
 
         /** Need to determine if it is the last criteria to be signed on the step */
-        const unsignedCriterias = workflowSteps
-            .find((x) => x.id === stepId)
-            ?.criterias.filter((x) => x.signedAtUtc === null);
+        const unsignedCriterias =
+            workflowSteps &&
+            workflowSteps
+                .find((x) => x.id === stepId)
+                ?.criterias.filter((x) => x.signedAtUtc === null);
 
         if (pendingContributions && unsignedCriterias?.length === 1) {
             spawnConfirmationDialog(