+++ src/components/TopBar/TopBar.tsx
@@ -1,9 +1,11 @@
-import { AddMenu } from '@equinor/DataFactory';
+// import { AddMenu } from '@equinor/DataFactory';
 import { Avatar, TopBar } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
 import { useClientContext } from '@equinor/lighthouse-portal-client';
 import { useRef, useState } from 'react';
 import { NotificationBell } from '../../Core/Notifications/Components/NotificationBell';
+import { AddMenu } from '../../FusionModules/DataCreatorReact/Components/AddMenu';
+import { useDataCreator } from '../../FusionModules/DataCreatorReact/Hooks/useCreator';
 import Icon from '../Icon/Icon';
 import { useMenuContext } from '../Menu';
 import { LocationBreadCrumbs } from './BreadCrumbs/Breadcrumbs';
@@ -15,6 +17,7 @@ import { Action, ActionWrapper, Header, Icons, TopBarWrapper } from './TopBarSty
 const ClientTopBar = (): JSX.Element => {
     // state for open and close add menu and add menu ref for positioning.
     const [isAddMenuOpen, setIsAddMenuOpen] = useState(false);
+    const { creators } = useDataCreator();
     const addMenuRef = useRef<HTMLHeadingElement>(null);
 
     const {
@@ -60,24 +63,26 @@ const ClientTopBar = (): JSX.Element => {
                     >
                         <HelpIcon />
                     </Action>
-                    <ActionWrapper ref={addMenuRef}>
-                        <Action
-                            title={'Add Item'}
-                            onFocus={() => setIsAddMenuOpen((s) => !s)}
-                            onMouseOver={() => {
-                                setIsAddMenuOpen(true);
-                            }}
-                            onBlur={() => setIsAddMenuOpen(false)}
-                        >
-                            <Icon name="add" />
-                        </Action>
-                        <AddMenu
-                            anchorEl={addMenuRef.current}
-                            isOpen={isAddMenuOpen}
-                            handleClose={() => setIsAddMenuOpen(false)}
-                            onMouseEnter={() => setIsAddMenuOpen(true)}
-                        />
-                    </ActionWrapper>
+                    {creators.length > 0 && (
+                        <ActionWrapper ref={addMenuRef}>
+                            <Action
+                                title={'Add Item'}
+                                onFocus={() => setIsAddMenuOpen((s) => !s)}
+                                onMouseOver={() => {
+                                    setIsAddMenuOpen(true);
+                                }}
+                                onBlur={() => setIsAddMenuOpen(false)}
+                            >
+                                <Icon name="add" />
+                            </Action>
+                            <AddMenu
+                                anchorEl={addMenuRef.current}
+                                isOpen={isAddMenuOpen}
+                                handleClose={() => setIsAddMenuOpen(false)}
+                                onMouseEnter={() => setIsAddMenuOpen(true)}
+                            />
+                        </ActionWrapper>
+                    )}
                     <Action
                         disabled
                         onClick={() => {