+++ src/packages/Sidesheet/Components/PopoutSidesheet.tsx
@@ -15,7 +15,7 @@ export const PopoutSidesheet = (): JSX.Element | null => {
      * Removes hash from url when closed
      */
     useEffect(() => {
-        if (location.hash.length > 0) return;
+        if (window.location.hash.length > 0) return;
         if (!props) {
             navigate(location.pathname, { replace: true });
         }