+++ Changelog.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 
 The changelog is valid starting with Castberg Project Portal v0.1.0-alpha.
+# 1.4.5
+- [MC] Production flag for analytics tab
+- [Portal] Change nginx version to fix dynatrace
+- [PipingAndHeatTrace] - 3D for tags
+
+## 1.4.4
+- [MC] Analytics tab
+- [Garden] Fix horizontal scroll
+
+## 1.4.3
+- [Portal] Access check hotfix
+- [MC] Now in production
+- [Table] Bugfixes
+
+
+## 1.4.2
+- [Scope change] Batch import tags and commPkgs
+- [Sidesheet] Sidesheet configuration
+- [Workspace] Workspace configuration
+- [Work order] Added KPI's
+- [Workspace] Small ui adjustments
+- [Workspace] Added support for help pages
+- [ModuleViewer] Possibility to turn on/off the rest of the 3D model
+
 
 ## 1.4.1
 - [PipingAndHeatTrace] Null-check bug fix. Updated app after FAM update.